Separated the chart settings widget
