+++ lib/src/oscilloscope/simple_oscilloscope.dart
@@ -1,5 +1,6 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:floscilloscope/src/oscilloscope/oscilloscope_axis_chart_data.dart';
+import 'package:floscilloscope/src/oscilloscope/settings/chart_settings.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:syncfusion_flutter_core/theme.dart';
@@ -18,38 +19,24 @@ class SimpleOscilloscope extends StatefulWidget {
 }
 
 class _SimpleOscilloscopeState extends State<SimpleOscilloscope> {
-  late final TextEditingController _horizontalAxisValuePerDivisionController;
-  late final TextEditingController _verticalAxisValuePerDivisionController;
-  late double _horizontalAxisValuePerDivision;
-  late double _verticalAxisValuePerDivision;
+
   double _thresholdProgressbarValue = 0.0;
   double _thresholdValue = 0.0;
   double _sliderBottomPadding = 0.0;
 
-  final FocusNode _horizontalAxisFocusNode = FocusNode();
-  final FocusNode _verticalAxisFocusNode = FocusNode();
   final GlobalKey<_SimpleOscilloscopeState> _chartAndLabelAreaRenderKey = GlobalKey<_SimpleOscilloscopeState>();
   final GlobalKey<_SimpleOscilloscopeState> _chartAreaRenderKey = GlobalKey<_SimpleOscilloscopeState>();
 
   @override
   void initState() {
     super.initState();
-    _horizontalAxisValuePerDivision = widget.oscilloscopeAxisChartData.defaultHorizontalAxisValuePerDivision;
-    _verticalAxisValuePerDivision = widget.oscilloscopeAxisChartData.defaultVerticalAxisValuePerDivision;
-    _horizontalAxisValuePerDivisionController = TextEditingController(text: _horizontalAxisValuePerDivision.toString());
-    _verticalAxisValuePerDivisionController = TextEditingController(text: _verticalAxisValuePerDivision.toString());
-
-    WidgetsBinding.instance!.addPostFrameCallback((_) {
+    WidgetsBinding.instance.addPostFrameCallback((_) {
       _calculateBottomPadding();
     });
   }
 
   @override
   void dispose() {
-    _horizontalAxisValuePerDivisionController.dispose();
-    _verticalAxisValuePerDivisionController.dispose();
-    _horizontalAxisFocusNode.dispose();
-    _verticalAxisFocusNode.dispose();
     super.dispose();
   }
 
@@ -63,12 +50,11 @@ class _SimpleOscilloscopeState extends State<SimpleOscilloscope> {
     }
   }
 
-  void _updateValuePerDivisions() {
-    _horizontalAxisValuePerDivision = double.tryParse(_horizontalAxisValuePerDivisionController.text) ?? _horizontalAxisValuePerDivision;
-    _verticalAxisValuePerDivision = double.tryParse(_verticalAxisValuePerDivisionController.text) ?? _verticalAxisValuePerDivision;
-    _horizontalAxisValuePerDivisionController.text = _horizontalAxisValuePerDivision.toString();
-    _verticalAxisValuePerDivisionController.text = _verticalAxisValuePerDivision.toString();
-    widget.oscilloscopeAxisChartData.onValuePerDivisionsChanged?.call(_horizontalAxisValuePerDivision, _verticalAxisValuePerDivision);
+  // This function will trigger the rebuild of the SimpleOscilloscope widget.
+  void _onSettingsUpdate() {
+    setState(() {
+      _updateThresholdValue();
+    });
   }
 
   void _updateThresholdValue([double? value]) {
@@ -78,44 +64,8 @@ class _SimpleOscilloscopeState extends State<SimpleOscilloscope> {
       _thresholdValue = value;
     }
     _thresholdProgressbarValue = _thresholdProgressbarValue.clamp(
-      -_verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
-      _verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
-    );
-  }
-
-  Widget _buildAxisInputField({
-    required TextEditingController controller,
-    required FocusNode focusNode,
-    required String label,
-    required String unit,
-  }) {
-    return Focus(
-      onFocusChange: (hasFocus) {
-        if (!hasFocus && controller.text.isEmpty) {
-          controller.text = focusNode == _horizontalAxisFocusNode
-              ? _horizontalAxisValuePerDivision.toString()
-              : _verticalAxisValuePerDivision.toString();
-        }
-      },
-      child: TextField(
-        focusNode: focusNode,
-        controller: controller,
-        decoration: InputDecoration(
-          labelText: '$label ($unit)',
-          border: const OutlineInputBorder(),
-        ),
-        keyboardType: const TextInputType.numberWithOptions(decimal: true),
-        inputFormatters: [
-          FilteringTextInputFormatter.allow(RegExp(r'^\d*\.?\d*')),
-        ],
-        onSubmitted: (value) {
-          setState(() {
-            _updateValuePerDivisions();
-            _updateThresholdValue();
-          });
-          focusNode.unfocus();
-        },
-      ),
+      -widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
+      widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
     );
   }
 
@@ -160,30 +110,30 @@ class _SimpleOscilloscopeState extends State<SimpleOscilloscope> {
                       gridData: FlGridData(
                         drawHorizontalLine: true,
                         drawVerticalLine: true,
-                        horizontalInterval: _verticalAxisValuePerDivision,
-                        verticalInterval: _horizontalAxisValuePerDivision,
+                        horizontalInterval: widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision,
+                        verticalInterval: widget.oscilloscopeAxisChartData.horizontalAxisValuePerDivision,
                         getDrawingHorizontalLine: (value) {
-                          return value % _verticalAxisValuePerDivision.abs() < _verticalAxisValuePerDivision
+                          return value % widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision.abs() < widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision
                               ? defaultGridLine(value)
                               : const FlLine(color: Colors.transparent);
                         },
                         getDrawingVerticalLine: (value) {
-                          return value % _horizontalAxisValuePerDivision.abs() < _horizontalAxisValuePerDivision
+                          return value % widget.oscilloscopeAxisChartData.horizontalAxisValuePerDivision.abs() < widget.oscilloscopeAxisChartData.horizontalAxisValuePerDivision
                               ? defaultGridLine(value)
                               : const FlLine(color: Colors.transparent);
                         },
                       ),
                       minX: 0,
-                      maxX: _horizontalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions * 2,
-                      minY: -_verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
-                      maxY: _verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
+                      maxX: widget.oscilloscopeAxisChartData.horizontalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions * 2,
+                      minY: -widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
+                      maxY: widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
                       clipData: const FlClipData.all(),
                       titlesData: FlTitlesData(
                         leftTitles: AxisTitles(
                           axisNameWidget: Text(widget.oscilloscopeAxisChartData.verticalAxisLabel),
                           axisNameSize: 18,
                           sideTitles: SideTitles(
-                            interval: _verticalAxisValuePerDivision,
+                            interval: widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision,
                             showTitles: true,
                             reservedSize: 85,
                             getTitlesWidget: (value, meta) {
@@ -200,7 +150,7 @@ class _SimpleOscilloscopeState extends State<SimpleOscilloscope> {
                           axisNameWidget: Text(widget.oscilloscopeAxisChartData.horizontalAxisLabel),
                           axisNameSize: 18,
                           sideTitles: SideTitles(
-                            interval: _horizontalAxisValuePerDivision,
+                            interval: widget.oscilloscopeAxisChartData.horizontalAxisValuePerDivision,
                             showTitles: true,
                             reservedSize: 45,
                             getTitlesWidget: (value, meta) {
@@ -276,8 +226,8 @@ class _SimpleOscilloscopeState extends State<SimpleOscilloscope> {
                               tooltipTextFormatterCallback: (dynamic actualValue, String formattedText) => _thresholdValue.toStringAsFixed(2),
                               overlayShape: const CustomOverlayShape(overlayRadius: 10),
                               thumbShape: const CustomThumbShape(thumbRadius: 10),
-                              min: -_verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
-                              max: _verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
+                              min: -widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
+                              max: widget.oscilloscopeAxisChartData.verticalAxisValuePerDivision * widget.oscilloscopeAxisChartData.numberOfDivisions,
                               value: _thresholdProgressbarValue,
                               showTicks: false,
                               showLabels: false,
@@ -308,42 +258,7 @@ class _SimpleOscilloscopeState extends State<SimpleOscilloscope> {
             child: SingleChildScrollView(
               child: Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    if (widget.oscilloscopeAxisChartData.settingsTitleLabel.isNotEmpty)
-                      Text(
-                        widget.oscilloscopeAxisChartData.settingsTitleLabel,
-                        style: const TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    const SizedBox(height: 20),
-                    _buildAxisInputField(
-                      controller: _horizontalAxisValuePerDivisionController,
-                      focusNode: _horizontalAxisFocusNode,
-                      label: widget.oscilloscopeAxisChartData.horizontalAxisTitlePerDivisionLabel,
-                      unit: widget.oscilloscopeAxisChartData.horizontalAxisUnit,
-                    ),
-                    const SizedBox(height: 20),
-                    _buildAxisInputField(
-                      controller: _verticalAxisValuePerDivisionController,
-                      focusNode: _verticalAxisFocusNode,
-                      label: widget.oscilloscopeAxisChartData.verticalAxisTitlePerDivisionLabel,
-                      unit: widget.oscilloscopeAxisChartData.verticalAxisUnit,
-                    ),
-                    const SizedBox(height: 20),
-                    if (widget.oscilloscopeAxisChartData.updateButtonLabel.isNotEmpty)
-                      ElevatedButton(
-                        onPressed: () => setState(() {
-                          _updateValuePerDivisions();
-                          _updateThresholdValue();
-                        }),
-                        child: Text(widget.oscilloscopeAxisChartData.updateButtonLabel),
-                      ),
-                  ],
-                ),
+                child: ChartSettings(oscilloscopeAxisChartData: widget.oscilloscopeAxisChartData, onSettingUpdateFunction: _onSettingsUpdate)
               ),
             ),
           ),
